Update index file and style header
Updated the index.html file to include a button in the hero section
as well as a line break in the hero section main text. Updated css file
and started styling the nav bar and the main hero section. Updates inclu
de background for header, logo styling, navigation link styling, and
hero styling on button, p tag, and h2.